refactor: Implements ContextApi instead of Redux
