+++ package.json
@@ -13,10 +13,8 @@
     "d3": "^6.5.0",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-redux": "^7.2.2",
     "react-scripts": "4.0.1",
-    "redux": "^4.0.5",
-    "redux-thunk": "^2.3.0",
+    "react-share": "^4.3.1",
     "victory": "^35.4.7",
     "web-vitals": "^0.2.4"
   },
